pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prometheus"
version = "0.1.0"
description = "Sovereign AI agent harness — extract, adapt, integrate"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "rich>=13.0",
    "anyio>=4.0",
    "python-telegram-bot>=21.0",
    "croniter>=2.0",
    "python-docx>=1.1.0",
    "PyMuPDF>=1.24.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
browser = ["playwright>=1.40"]
mcp = ["mcp>=1.0"]
slack = ["slack-bolt>=1.18.0", "slack-sdk>=3.27.0"]
evals = [
    "deepeval>=2.0",
    "arize-phoenix>=5.0",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[project.scripts]
prometheus = "prometheus.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/prometheus"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = ["integration: integration wiring tests (verify components are connected at runtime)"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]
